Pmd.codedeploy.2016.01.06

+++ bin/codepipeline-create.sh
@@ -53,8 +53,8 @@ generate_cli_json() {
 _END_
 }
 
-aws codepipeline create-custom-action-type --cli-input-json "$(generate_cli_json Build)"
-aws codepipeline create-custom-action-type --cli-input-json "$(generate_cli_json Test)"
+# aws codepipeline create-custom-action-type --cli-input-json "$(generate_cli_json Build)"
+# aws codepipeline create-custom-action-type --cli-input-json "$(generate_cli_json Test)"
 
 pipelinejson=$(mktemp /tmp/dromedary-pipeline.json.XXXX)
 pipelinedeployjson=$(mktemp /tmp/dromedary-pipeline-deploy.json.XXXX)
@@ -85,11 +85,27 @@ echo The value of variable dromedary_pipeline_codedeploy_stack_name = $dromedary
 echo The value of variable dromedary_codedeploy_config_name = $dromedary_codedeploy_config_name 
 echo The value of variable dromedary_codedeploy_app_name = $dromedary_codedeploy_app_name 
 
-aws cloudformation create-stack --stack-name $dromedary_pipeline_stack_name --template-body file://$pipelinejson --region us-east-1 --disable-rollback --capabilities="CAPABILITY_IAM" --parameters ParameterKey=GitHubToken,ParameterValue=$mygithubtoken ParameterKey=GitHubUser,ParameterValue=$mygithubuser ParameterKey=Branch,ParameterValue=$mybranch
-aws cloudformation create-stack --stack-name $dromedary_pipeline_codedeploy_stack_name --template-body file://$pipelinedeployjson --region us-east-1 --disable-rollback --capabilities="CAPABILITY_IAM" --parameters ParameterKey=GitHubToken,ParameterValue=$mygithubtoken ParameterKey=GitHubUser,ParameterValue=$mygithubuser ParameterKey=Branch,ParameterValue=$mybranch ParameterKey=MyDeploymentConfigName,ParameterValue=$dromedary_codedeploy_config_name ParameterKey=MyApplicationName,ParameterValue=$dromedary_codedeploy_app_name
+# Create Custom Actions
+aws cloudformation create-stack \
+    --stack-name $dromedary_pipeline_customactions_stack_name \
+    --capabilities CAPABILITY_IAM \
+    --template-body file://./pipeline/cfn/codepipeline-custom-actions.json \
+    --parameters ParameterKey=MyBuildProvider,ParameterValue=$dromedary_custom_action_provider ParameterKey=MyJenkinsURL,ParameterValue=$jenkins_url  
+
+customactions_stack_status="$(bash $script_dir/cfn-wait-for-stack.sh $dromedary_pipeline_customactions_stack_name)"
+customactions_stack_wait=$?
+
+echo
+
+if [ $customactions_stack_wait -ne 0 ]; then
+    echo "Fatal: Custom Actions stack $dromedary_pipeline_customactions_stack_name ($customactions_stack_status) failed to create properly" >&2
+    exit 1
+fi
+
+aws cloudformation create-stack --stack-name $dromedary_pipeline_stack_name --template-body file://$pipelinejson --region us-east-1 --disable-rollback --capabilities="CAPABILITY_IAM" --parameters ParameterKey=GitHubToken,ParameterValue=$mygithubtoken ParameterKey=GitHubUser,ParameterValue=$mygithubuser ParameterKey=Branch,ParameterValue=$mybranch ParameterKey=MyJenkinsURL,ParameterValue=$jenkins_url ParameterKey=MyBuildProvider,ParameterValue=$dromedary_custom_action_provider 
+# aws cloudformation create-stack --stack-name $dromedary_pipeline_codedeploy_stack_name --template-body file://$pipelinedeployjson --region us-east-1 --disable-rollback --capabilities="CAPABILITY_IAM" --parameters ParameterKey=GitHubToken,ParameterValue=$mygithubtoken ParameterKey=GitHubUser,ParameterValue=$mygithubuser ParameterKey=Branch,ParameterValue=$mybranch ParameterKey=MyDeploymentConfigName,ParameterValue=$dromedary_codedeploy_config_name ParameterKey=MyApplicationName,ParameterValue=$dromedary_codedeploy_app_name ParameterKey=MyBuildProvider,ParameterValue=$dromedary_custom_action_provider
 
 echo "export dromedary_codepipeline=$dromedary_pipeline_stack_name" >> "$ENVIRONMENT_FILE"
 echo "export dromedary_codepipeline_codedeploy=$dromedary_pipeline_codedeploy_stack_name" >> "$ENVIRONMENT_FILE"
 rm -f $pipelinejson
 rm -f $pipelinedeployjson
-
